requirements.txt
anthropic>=0.40.0
streamlit>=1.29.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
fpdf2>=2.7.0
